move _sturges_formula
